backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-firestore==2.16.0
google-cloud-bigquery==3.25.0
